Setting royalties as per EIP2981 / ERC2981
a small tweak to implement EIP2981 to set royalties.
I understand it is up to the marketplace to honour it (e.g. OpenSea and Rarible have their own "hack" for royalties) but it seems the EIP2981 is starting to be adopted by newer marketplaces and hopefully the dinosaurs will eventually follow. 
these changes are small and hopefully make this contract future proof

as per Liarco suggestion, I've created a way to set royalties at any time, but have a default open in the construct

*note: I've only used this in a couple of test contracts, so please do your own tests and submit any feedback  and/or tweaks here